requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-image==0.21.0
matplotlib==3.7.2
Pillow==10.0.0
scipy==1.11.3